fix: int -> Long 으로 수정

+++ src/main/java/com/collbus/assignment/domain/Posts.java
@@ -24,5 +24,5 @@ public class Posts {
     private Long createdUserId;
 
     @Column(name = "likes_count")
-    private int likesCount;
+    private Long likesCount;
 }
